Move independent java function to separate module
