+++ bin/commands/internal/start/prepare/index.ts
@@ -24,6 +24,7 @@ import * as config from '../../../../libs/config';
 import * as component from '../../../../libs/component';
 import * as varlib from '../../../../libs/var';
 import * as java from '../../../../libs/java';
+import * as javaCI from '../../../../libs/java_ci';
 import * as node from '../../../../libs/node';
 import * as zosmf from '../../../../libs/zosmf';
 
@@ -148,7 +149,7 @@ function globalValidate(enabledComponents:string[]): void {
     // validate java for some core components
     //TODO this should be a manifest parameter that you require java, not a hardcoded list. What if extensions require it?
     if (enabledComponents.includes('gateway') || enabledComponents.includes('zaas') || enabledComponents.includes('discovery') || enabledComponents.includes('api-catalog') || enabledComponents.includes('caching-service')) {
-      let javaOk = java.validateJavaHome();
+      let javaOk = javaCI.validateJavaHome();
       if (!javaOk) {
         privateErrors++;
         common.printFormattedError('ZWELS', "zwe-internal-start-prepare,global_validate", `Could not validate java home`);
