+++ bin/libs/java.ts
@@ -13,10 +13,9 @@ import * as std from 'cm_std';
 import * as os from 'cm_os';
 import * as fs from './fs';
 import * as common from './common';
+import * as javaCI from './java_ci'
 import * as shell from './shell';
 import * as config from './config';
-
-const JAVA_MIN_VERSION=17;
 
 export function ensureJavaIsOnPath(): void {
   let path=std.getenv('PATH') || '/bin:.:/usr/bin';
@@ -30,7 +29,7 @@ export function readConfigJavaHome(configList?: string, skipValidate?: boolean):
   const zoweConfig = config.getZoweConfig();
   if (zoweConfig && zoweConfig.java && zoweConfig.java.home) {
     if (!skipValidate) {
-      if (!validateJavaHome(zoweConfig.java.home)) {
+      if (!javaCI.validateJavaHome(zoweConfig.java.home)) {
         return '';
       }
     }
@@ -77,66 +76,6 @@ export function requireJava() {
 
   ensureJavaIsOnPath();
   _javaCheckComplete = true;
-}
-
-export function validateJavaHome(javaHome:string|undefined=std.getenv("JAVA_HOME")): boolean {
-  if (!javaHome) {
-    common.printError("Cannot find java. Please define JAVA_HOME environment variable.");
-    return false;
-  }
-  if (!fs.fileExists(fs.resolvePath(javaHome,`/bin/java`))) {
-    common.printError(`JAVA_HOME: ${javaHome}/bin does not point to a valid install of Java.`);
-    return false;
-  }
-
-  let execReturn = shell.execErrSync(fs.resolvePath(javaHome,`/bin/java`), `-version`);
-  const version = execReturn.err;
-  if (execReturn.rc != 0) {
-    common.printError(`Java version check failed with return code: ${execReturn.rc}: ${version}`);
-    return false;
-  }
- 
-  try {
-    let index = 0;
-    let javaVersionShort;
-    let versionLines = (version as string).split('\n'); // valid because of above rc check
-    for (let i = 0; i < versionLines.length; i++) {
-      if ((index = versionLines[i].indexOf('java version')) != -1) {
-        //format of: java version "1.8.0_321"
-        javaVersionShort=versionLines[i].substring(index+('java version'.length)+2, versionLines[i].length-1);
-        break;
-      } else if ((index = versionLines[i].indexOf('openjdk version')) != -1) {
-        javaVersionShort=versionLines[i].substring(index+('openjdk version'.length)+2, versionLines[i].length-1);
-        break;
-      }
-    }
-    if (!javaVersionShort){
-      common.printError("could not find java version");
-      return false;
-    }
-    let versionParts = javaVersionShort.split('.');
-    const javaMajorVersion=Number(versionParts[0]);
-    const javaMinorVersion=Number(versionParts[1]);
-
-    let tooLow=false;
-    if (javaMajorVersion !== 1 && javaMajorVersion < JAVA_MIN_VERSION) {
-      tooLow=true;
-    }
-    if (javaMajorVersion === 1 && javaMinorVersion < JAVA_MIN_VERSION) {
-      tooLow=true;
-    }
-
-    if (tooLow) {
-      common.printError(`Java ${javaVersionShort} is less than the minimum level required of Java ${JAVA_MIN_VERSION}.`);
-      return false;
-    }
-
-    common.printDebug(`Java ${javaVersionShort} is supported.`);
-    common.printDebug(`Java check is successful.`);
-    return true;
-  } catch (e) {
-    return false;
-  }
 }
 
 export function getJavaPkcs12KeystoreFlag(javaHome:string|undefined=std.getenv("JAVA_HOME")): string {
